Snowflake destination: add support for ecrypted private key
